Arrow and Spell and Buff System redone
they all tie in with the item system.
they all hold the effect on the respectvie scriptable object
intuitive chances , ex: create a new arrow by creating a new item, a new arrow , setting the item.arrow slot with the arrow. the prefab sits on the arrow SO.
it all works