gold commint
gold system in place
first work on the player spawning system was done in this commit. the game should work in the end without a player. still needs a lot of work
this lays the ground for the player prefs implementation and the setup scene where there will be stats distribution, class select and so on